+++ example/counter/web/src/components/CounterDisplay.dart
@@ -1,0 +1,19 @@
+import 'dart:html';
+
+import 'package:wire/wire.dart';
+
+import '../const/CounterParams.dart';
+import 'base/dom_element.dart';
+
+class CounterDisplay extends DomElement {
+  CounterDisplay():super(DivElement()) {
+    dom.className = 'spectrum-Body spectrum-Body--L';
+
+    final wireData = Wire.data(CounterDataKeys.COUNT);
+    wireData.subscribe(update);
+    update(wireData.value ?? 0);
+  }
+
+  void update(value) { dom.text = value.toString(); }
+}
+