+++ lib/src/const.dart
@@ -2,4 +2,5 @@ part of wire;
 
 final String ERROR__MIDDLEWARE_EXISTS =
     'Middleware already registered, middleware: ';
-final String ERROR__WIRE_ALREADY_REGISTERED = 'Wire already registered, wid: ';
+final String ERROR__WIRE_ALREADY_REGISTERED =
+    'Wire already registered, wid: ';