+++ lib/src/main.dart
@@ -14,7 +14,7 @@ part 'data.dart';
 /// License: APACHE LICENSE, VERSION 2.0
 ///
 
-typedef WireListener<T> = void Function(T payload, int wid);
+typedef WireListener<T> = void Function(T payload, int wireId);
 
 abstract class WireMiddleware {
   void onAdd(Wire wire);
@@ -61,8 +61,8 @@ class Wire<T> {
   /// Unique identification for wire instance.
   ///
   /// @private
-  int _wid;
-  int get wid => _wid;
+  int _id;
+  int get id => _id;
 
   ///
   /// [read-only] [internal use]
@@ -84,21 +84,29 @@ class Wire<T> {
   /// But it wont react on signal until it is attached to the communication layer with [attach]
   /// However you still can send data through it by calling [transfer]
   ///
-  Wire(Object scope, String signal, WireListener<T> listener, [int replies = 0])
+  Wire(
+    Object scope,
+    String signal,
+    WireListener<T> listener,
+    [int replies = 0]
+  )
       : assert(scope != null),
         assert(signal != null),
-        assert(listener != null) {
+        assert(listener != null)
+  {
     _scope = scope;
     _signal = signal;
     _listener = listener;
     this.replies = replies;
-    _wid = ++_INDEX;
+    _id = ++_INDEX;
   }
 
   /// Call associated WireListener with data.
-  void transfer(payload) {
+  void transfer(dynamic payload) {
     // Call a listener in this Wire only in case data type match it's listener type.
-    if (payload is T) _listener(payload, _wid);
+    if (payload is T || payload == null) {
+      _listener(payload, _id);
+    }
   }
 
   /// Nullify all relations
@@ -121,13 +129,18 @@ class Wire<T> {
 
   /// Remove wire object from communication layer, returns existence.
   static bool detach(Wire wire) {
-    return _COMMUNICATION_LAYER.remove(wire.signal, wire.scope, wire.listener);
+    return _COMMUNICATION_LAYER.remove(
+        wire.signal, wire.scope, wire.listener);
   }
 
   /// Create wire object from params and [attach] it to the communication layer
   /// All middleware will be informed from [WireMiddleware.onAdd] before wire is attached to the layer
-  static Wire add<T>(Object scope, String signal, WireListener<T> listener,
-      {int replies = 0}) {
+  static Wire add<T>(
+    Object scope,
+    String signal,
+    WireListener<T> listener,
+    {int replies = 0}
+  ) {
     final wire = Wire<T>(scope, signal, listener, replies);
     _MIDDLEWARE_LIST.forEach((m) => m.onAdd(wire));
     attach(wire);
@@ -135,7 +148,12 @@ class Wire<T> {
   }
 
   /// Check if signal string or wire instance exists in communication layer
-  static bool has({String signal, Wire wire}) {
+  static bool has(
+    {
+      String signal,
+      Wire wire
+    }
+  ) {
     if (signal != null) return _COMMUNICATION_LAYER.hasSignal(signal);
     if (wire != null) return _COMMUNICATION_LAYER.hasWire(wire);
     return false;
@@ -146,14 +164,24 @@ class Wire<T> {
   /// If use scope then only wire with this scope value will receive the payload
   /// All middleware will be informed from [WireMiddleware.onSend] before signal sent on wires
   /// Returns true when no wire for the signal has found
-  static bool send<T>(String signal, {T payload, Object scope}) {
+  static bool send<T>(
+    String signal,
+    {
+      T payload,
+      Object scope
+    }
+  ) {
     _MIDDLEWARE_LIST.forEach((m) => m.onSend(signal, payload));
     return _COMMUNICATION_LAYER.send(signal, payload, scope);
   }
 
   /// Remove all entities from Communication Layer and Data Container Layer
   /// @param [withMiddleware] used to remove all middleware
-  static void purge({bool withMiddleware}) {
+  static void purge(
+    {
+      bool withMiddleware
+    }
+  ) {
     _COMMUNICATION_LAYER.clear();
     _DATA_CONTAINER_LAYER.clear();
     if (withMiddleware ?? false) _MIDDLEWARE_LIST.clear();
@@ -162,7 +190,13 @@ class Wire<T> {
   /// Remove all wires for specific signal, for more precise target to remove add scope and/or listener
   /// All middleware will be informed from [WireMiddleware.onRemove] after signal removed, only if existed
   /// Returns [bool] telling signal existed in communication layer
-  static bool remove(String signal, {Object scope, WireListener listener}) {
+  static bool remove(
+    String signal,
+    {
+      Object scope,
+      WireListener listener
+    }
+  ) {
     var existed = _COMMUNICATION_LAYER.remove(signal, scope, listener);
     if (existed) {
       _MIDDLEWARE_LIST.forEach((m) => m.onRemove(signal, scope, listener));
@@ -182,7 +216,13 @@ class Wire<T> {
   /// When you need Wires associated with signal or scope or listener
   /// Returns [List<Wire>]
   static List<Wire> get(
-      {String signal, Object scope, WireListener listener, int wid}) {
+    {
+      String signal,
+      Object scope,
+      WireListener listener,
+      int wireId
+    }
+  ) {
     var result = <Wire>[];
     if (signal != null) {
       result.addAll(_COMMUNICATION_LAYER.getBySignal(signal));
@@ -193,8 +233,8 @@ class Wire<T> {
     if (listener != null) {
       result.addAll(_COMMUNICATION_LAYER.getByListener(listener));
     }
-    if (wid != null) {
-      result.add(_COMMUNICATION_LAYER.getByWID(wid));
+    if (wireId != null) {
+      result.add(_COMMUNICATION_LAYER.getByWID(wireId));
     }
     return result;
   }
@@ -212,7 +252,12 @@ class Wire<T> {
   /// void remove()
   /// ```
   /// Returns [WireData]
-  static WireData data<T>(String key, [T value]) {
+  static WireData data<T>(
+    String key,
+    [
+      T value
+    ]
+  ) {
     var wireData = _DATA_CONTAINER_LAYER.get(key);
     if (value != null) {
       var prevValue = wireData.value;